restructure the controller and services
