--- backend/ISettlyService/IEmailVerificationService.cs
@@ -1,6 +1,0 @@
-namespace ISettlyService;
-
-public interface IEmailVerificationService
-{
-    Task GenerateAndSendVerificationCodeAsync(int userId, string email);
-}
