+++ backend/ISettlyService/IVerificationCodeService.cs
@@ -1,0 +1,6 @@
+namespace ISettlyService;
+
+public interface IVerificationCodeService
+{
+    Task<string> GenerateAndSaveCodeAsync(int userId);
+}
